requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pytesseract>=0.3.10

# ML/DL
torch>=2.0.0
transformers>=4.35.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0

# Utils
pydantic>=2.0.0
python-dotenv>=1.0.0
jsonschema>=4.19.0
Faker>=20.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
